Corrected bug in WinRTDelete
The SQLite documentation suggests that the delete function may return
before changes are committed to disk. In practice, however, for
journaling, the open function may be called to create a journal file
immediately after it has been deleted. If a delete operation is pending
when this occurs, the VFS will crash. Therefore, we must ALWAYS await
the delete operation before continuing.